Refactor UI module with modular components and improved interactive runner
